+++ CHANGELOG.md
@@ -44,6 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix race conditions in SQLite database test runner [#345](https://github.com/p2panda/aquadoggo/pull/345)
 - Remove duplicate `Cursor` implementation, query code clean up [#346](https://github.com/p2panda/aquadoggo/pull/346)
 - Fix SQL ordering to correctly assemble resulting rows to documents [#347](https://github.com/p2panda/aquadoggo/pull/347)
+- Fix parsing lists arguments twice in GraphQL [#354](https://github.com/p2panda/aquadoggo/pull/354)
+- Sort paginated query field rows by document view id [#354](https://github.com/p2panda/aquadoggo/pull/354)
 
 ## [0.4.0]